JPC: Authorize-form refactor
