+++ assets/stylesheets/_components.scss
@@ -438,6 +438,8 @@
 @import 'signup/style';
 @import 'signup/flow-progress-indicator/style';
 @import 'signup/jetpack-connect/style';
+@import 'signup/jetpack-connect/authorize-form/style';
+@import 'signup/jetpack-connect/jetpack-site-card/style';
 @import 'signup/jetpack-connect/plan-features-tab/style';
 @import 'signup/locale-suggestions/style';
 @import 'signup/navigation-link/style';
