+++ client/signup/jetpack-connect/authorize-form/style.scss
@@ -1,0 +1,38 @@
+.jetpack-connect__authorize-form {
+	.jetpack-connect__authorize-form-header {
+		text-align: center;
+	}
+}
+
+.jetpack-connect__logged-in-form {
+	.jetpack-connect__logged-in-form-user-text {
+		text-align: center;
+	}
+
+	.gravatar {
+		display: block;
+		margin: 0 auto 8px auto;
+	}
+
+	.button {
+		width: 100%;
+	}
+
+	.logged-out-form__links .gridicon {
+		top: 2px;
+	}
+}
+
+.jetpack-connect__logged-in-form-loading {
+	text-align: center;
+
+	span {
+		animation: loading-fade 1.6s ease-in-out infinite;
+		display: block;
+	}
+
+	.spinner {
+		display: inline-block;
+		margin-top: 8px;
+	}
+}
