+++ client/signup/jetpack-connect/jetpack-site-card/style.scss
@@ -1,0 +1,3 @@
+.jetpack-connect__site.card {
+	padding: 0;
+}
